data_lakehouse/requirements.txt
delta-spark==1.0.1
lxml
python-decouple
requests
pandas
great-expectations
boto3
psycopg2-binary
